gym_saas/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
psycopg2-binary
Flask-Bcrypt==1.0.1
gunicorn